refactor: mv {init,bundle,serve} to src/subcommands
