--- examples/serve-command-example/deno.json
@@ -1,1 +1,0 @@
-{}
