+++ src/subcommands/init.ts
@@ -29,12 +29,12 @@ export async function main(
   switch (options.template) {
     case "serve": {
       await copyFile({
-        from: "./examples/serve-command-example/client.tsx",
+        from: "./templates/serve/client.tsx",
         to: "./client.tsx",
         debug: options.debug,
       });
       await copyFile({
-        from: "./examples/serve-command-example/app.ts",
+        from: "./templates/serve/app.ts",
         to: "./app.ts",
         debug: options.debug,
       });
